src/main/java/stc/inno20/dao/MobileDaoJdbcImpl.java
package stc.inno20.dao;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import stc.inno20.ConnectionManager.ConnectionManager;
import stc.inno20.pojo.Mobile;

import java.sql.Connection;
import java.sql.PreparedStatement;
import java.sql.ResultSet;
import java.sql.SQLException;

public class MobileDaoJdbcImpl implements MobileDao {
  private static final Logger LOGGER = LoggerFactory.getLogger(MobileDaoJdbcImpl.class);
  public static final String INSERT_INTO_MOBILE = "INSERT INTO mobile values (DEFAULT, ?, ?, ?)";
  public static final String SELECT_FROM_MOBILE = "SELECT * FROM mobile WHERE id = ?";
  public static final String UPDATE_MOBILE = "UPDATE mobile SET model=?, price=?, manufacturer=? WHERE id=?";
  public static final String DELETE_FROM_MOBILE = "DELETE FROM mobile WHERE id=?";


  private ConnectionManager connectionManager;

  public MobileDaoJdbcImpl(ConnectionManager connectionManager) {
    this.connectionManager = connectionManager;
  }

  @Override
  public boolean addMobile(Mobile mobile) {
    try (Connection connection = connectionManager.getConnection();
        PreparedStatement preparedStatement = connection.prepareStatement(INSERT_INTO_MOBILE)) {
      preparedStatement.setString(1, mobile.getModel());
      preparedStatement.setInt(2, mobile.getPrice());
      preparedStatement.setString(3, mobile.getManufacturer());
      preparedStatement.execute();
    } catch (SQLException e) {
      LOGGER.error("Some thing wrong in addMobile method", e);
      return false;
    }
    return true;
  }

  @Override
  public Mobile getMobileById(Integer id) {
    try (Connection connection = connectionManager.getConnection();
        PreparedStatement preparedStatement = connection.prepareStatement(SELECT_FROM_MOBILE)) {
      preparedStatement.setInt(1, id);
      try (ResultSet resultSet = preparedStatement.executeQuery()) {
        if (resultSet.next()) {
          return new Mobile(
              resultSet.getInt(1),
              resultSet.getString(2),
              resultSet.getInt(3),
              resultSet.getString(4));
        }
      }
    } catch (SQLException e) {
      LOGGER.error("Some thing wrong in getMobileById method", e);
    }
    return null;
  }

  @Override
  public boolean updateMobileById(Mobile mobile) {

    try (Connection connection = connectionManager.getConnection();
        PreparedStatement preparedStatement = connection.prepareStatement(UPDATE_MOBILE)) {
      preparedStatement.setString(1, mobile.getModel());
      preparedStatement.setInt(2, mobile.getPrice());
      preparedStatement.setString(3, mobile.getManufacturer());
      preparedStatement.setInt(4, mobile.getId());
      preparedStatement.execute();
      return true;
    } catch (SQLException e) {
      LOGGER.error("Some thing wrong in updateMobileById method", e);
    }
    return false;
  }

  @Override
  public boolean deleteMobileById(Integer id) {
    try (Connection connection = connectionManager.getConnection();
        PreparedStatement preparedStatement = connection.prepareStatement(DELETE_FROM_MOBILE)) {
      preparedStatement.setInt(1, id);
      preparedStatement.execute();
    } catch (SQLException e) {
      LOGGER.error("Some thing wrong in deleteMobileById method", e);
      return false;
    }
    return true;
  }
}
